refactor: remove formatError, convertToSignableToken
formatError, and convertToSignableToken methods are only used internally and don't need to be exported

[DX-1251]
